[FEATURE] Apply and use RFC 4155 compliant mbox mail transport
To split mails without boundary correctly.
Also, the armin/mbox-parser library has been updated to 1.1.0.
Thanks to Philipp Thiele for reporting this issue.

Also, do not show tabs for html or text, when this part is not given.
And do not throw exception, when path to mbox is wrong.
Instead, show error message as card in module.